backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Data loading
h5py>=3.10.0
numpy>=1.26.0
datasets>=2.16.0
huggingface_hub>=0.20.0
webdataset>=0.2.86
mcap>=1.1.0
mcap-protobuf-support>=0.5.0
Pillow>=10.0.0

# Quality metrics
scikit-learn>=1.4.0
scipy>=1.12.0
opencv-python>=4.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0
